+++ t/EditSections.t
@@ -37,7 +37,7 @@ require App::Followme::EditSections;
 
 my $test_dir = catdir(@path, 'test');
 
-rmtree($test_dir);
+rmtree($test_dir) if -e $test_dir;
 mkdir $test_dir  or die $!;
 chmod 0755, $test_dir;
