+++ t/FileData.t
@@ -23,7 +23,7 @@ eval "use App::Followme::FIO";
 require App::Followme::FileData;
 
 my $test_dir = catdir(@path, 'test');
-rmtree($test_dir);
+rmtree($test_dir) if -e $test_dir;
 
 mkdir $test_dir or die $!;
 chmod 0755, $test_dir;
