+++ t/FormatPage.t
@@ -5,7 +5,7 @@ use IO::File;
 use File::Path qw(rmtree);
 use File::Spec::Functions qw(catdir catfile rel2abs splitdir);
 
-use Test::More tests => 29;
+use Test::More tests => 33;
 
 #----------------------------------------------------------------------
 # Change the modification date of a file
@@ -38,7 +38,7 @@ require App::Followme::FormatPage;
 
 my $test_dir = catdir(@path, 'test');
 
-rmtree($test_dir);
+rmtree($test_dir) if -e $test_dir;
 mkdir $test_dir or die $!;
 chmod 0755, $test_dir;
 
@@ -294,23 +294,25 @@ do {
             my $filename = catfile($path, "$count.html");
             my $input = fio_read_page($filename);
 
+            ok(length($input) > 0, "Read $filename"); # test 14, 19, 24, 29
+
             like($input, qr(Page $count),
-               "Format block in $dir/$count"); # test 14, 18, 22, 26
+               "Format block in $dir/$count"); # test 15, 20, 25, 30
 
             like($input, qr(top link),
-               "Format prototype $dir/$count"); # test 15, 19, 23 27
+               "Format prototype $dir/$count"); # test 16, 21, 26 31
 
             if ($dir) {
                 like($input, qr(section nav in sub --),
-                   "Format section tag in $dir/$count"); # test 24, 28
+                   "Format section tag in $dir/$count"); # test 17, 22
                 like($input, qr(link one),
-                   "Format folder block $dir/$count"); # test 25, 29
+                   "Format folder block $dir/$count"); # test 18, 23
 
             } else {
                 like($input, qr(section nav --),
-                   "Format section tag in $dir/$count"); # test 16, 20
+                   "Format section tag in $dir/$count"); # test 27, 32
                 like($input, qr(link $count),
-                   "Format folder block in $dir/$count"); # test 17, 23
+                   "Format folder block in $dir/$count"); # test 28, 33
             }
         }
     }
